refactor: restructure game processing and enhance replay functionality
- Replaced GameParser with GameReplayer in core processing to streamline game analysis and replay.
- Introduced GameAssembler for improved PGN line parsing and game object assembly.
- Updated tracking and replay logic to utilize new SAN handling classes (SanApplier and SanToActions).
- Enhanced movetext processing by adding a dedicated movetext-tokenizer module for better clarity and performance.
- Updated AGENTS.md and IDEAS.md to reflect changes in architecture and future enhancements.
- Adjusted package.json imports to align with new module structure.
